Renamed attribute, moved it to Core namespace
