+++ src/Serilog/Core/Pipeline/SilentLogger.cs
@@ -44,12 +44,12 @@ namespace Serilog.Core.Pipeline
         {
         }
         
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Write(LogEventLevel level, string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Write(LogEventLevel level, Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
@@ -59,52 +59,52 @@ namespace Serilog.Core.Pipeline
             return false;
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Verbose(string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Verbose(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Debug(string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Debug(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Information(string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Information(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Warning(string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Warning(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Error(string messageTemplate, params object[] propertyValues)
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Error(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
@@ -113,7 +113,7 @@ namespace Serilog.Core.Pipeline
         {
         }
 
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         public void Fatal(Exception exception, string messageTemplate, params object[] propertyValues)
         {
         }
