+++ src/Serilog/ILogger.cs
@@ -77,7 +77,7 @@ namespace Serilog
         /// <param name="level">The level of the event.</param>
         /// <param name="messageTemplate"></param>
         /// <param name="propertyValues"></param>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Write(LogEventLevel level, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -87,7 +87,7 @@ namespace Serilog
         /// <param name="exception">Exception related to the event.</param>
         /// <param name="messageTemplate">Message template describing the event.</param>
         /// <param name="propertyValues">Objects positionally formatted into the message template.</param>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Write(LogEventLevel level, Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -106,7 +106,7 @@ namespace Serilog
         /// <example>
         /// Log.Verbose("Staring into space, wondering if we're alone.");
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Verbose(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -118,7 +118,7 @@ namespace Serilog
         /// <example>
         /// Log.Verbose(ex, "Staring into space, wondering where this comet came from.");
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Verbose(Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -129,7 +129,7 @@ namespace Serilog
         /// <example>
         /// Log.Debug("Starting up at {StartedAt}.", DateTime.Now);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Debug(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -141,7 +141,7 @@ namespace Serilog
         /// <example>
         /// Log.Debug(ex, "Swallowing a mundane exception.");
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Debug(Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -152,7 +152,7 @@ namespace Serilog
         /// <example>
         /// Log.Information("Processed {RecordCount} records in {TimeMS}.", records.Length, sw.ElapsedMilliseconds);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Information(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -164,7 +164,7 @@ namespace Serilog
         /// <example>
         /// Log.Information(ex, "Processed {RecordCount} records in {TimeMS}.", records.Length, sw.ElapsedMilliseconds);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Information(Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -175,7 +175,7 @@ namespace Serilog
         /// <example>
         /// Log.Warning("Skipped {SkipCount} records.", skippedRecords.Length);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Warning(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -187,7 +187,7 @@ namespace Serilog
         /// <example>
         /// Log.Warning(ex, "Skipped {SkipCount} records.", skippedRecords.Length);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Warning(Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -198,7 +198,7 @@ namespace Serilog
         /// <example>
         /// Log.Error("Failed {ErrorCount} records.", brokenRecords.Length);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Error(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -210,7 +210,7 @@ namespace Serilog
         /// <example>
         /// Log.Error(ex, "Failed {ErrorCount} records.", brokenRecords.Length);
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Error(Exception exception, string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -221,7 +221,7 @@ namespace Serilog
         /// <example>
         /// Log.Fatal("Process terminating.");
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Fatal(string messageTemplate, params object[] propertyValues);
 
         /// <summary>
@@ -233,7 +233,7 @@ namespace Serilog
         /// <example>
         /// Log.Fatal(ex, "Process terminating.");
         /// </example>
-        [LoggerMethod("messageTemplate")]
+        [MessageTemplateFormatMethod("messageTemplate")]
         void Fatal(Exception exception, string messageTemplate, params object[] propertyValues);
     }
 }
