+++ src/Serilog/Serilog-net40.csproj
@@ -49,6 +49,7 @@
     <Compile Include="Core\Enrichers\PropertyEnricher.cs" />
     <Compile Include="Core\IScalarConversionPolicy.cs" />
     <Compile Include="Core\LoggingLevelSwitch.cs" />
+    <Compile Include="Core\MessageTemplateFormatMethodAttribute.cs" />
     <Compile Include="Core\Sinks\CopyingSink.cs" />
     <Compile Include="Core\Filters\DelegateFilter.cs" />
     <Compile Include="Core\Sinks\FilteringSink.cs" />
@@ -62,7 +63,6 @@
     <Compile Include="Formatting\Display\LiteralStringValue.cs" />
     <Compile Include="Formatting\Display\MessageTemplateTextFormatter.cs" />
     <Compile Include="Formatting\Raw\RawFormatter.cs" />
-    <Compile Include="LoggerMethodAttribute.cs" />
     <Compile Include="Parameters\DepthLimiter.cs" />
     <Compile Include="Parameters\MessageTemplateProcessor.cs" />
     <Compile Include="Core\Sinks\SafeAggregateSink.cs" />
