+++ src/Serilog/Serilog.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Core\Constants.cs" />
     <Compile Include="Core\Enrichers\PropertyEnricher.cs" />
     <Compile Include="Core\IScalarConversionPolicy.cs" />
+    <Compile Include="Core\MessageTemplateFormatMethodAttribute.cs" />
     <Compile Include="Core\Sinks\CopyingSink.cs" />
     <Compile Include="Core\Filters\DelegateFilter.cs" />
     <Compile Include="Core\Sinks\FilteringSink.cs" />
@@ -67,7 +68,6 @@
     <Compile Include="Filters\Matching.cs" />
     <Compile Include="Formatting\Display\LiteralStringValue.cs" />
     <Compile Include="Formatting\Raw\RawFormatter.cs" />
-    <Compile Include="LoggerMethodAttribute.cs" />
     <Compile Include="Parameters\DepthLimiter.cs" />
     <Compile Include="Parameters\GetablePropertyFinder.cs" />
     <Compile Include="Parameters\MessageTemplateProcessor.cs" />
